wasm-jit: data reconciliation (-reconcile) (#16424)
Port `SimulationRuntime/c/dataReconciliation/dataReconciliation.cpp`
(VDI 2048 D.1, D.2 and state estimation) to the wasm-jit runtime, which
until now rejected the flag outright:

    LOG_ERROR | error | wasm-jit simulation failed:
                        -reconcile: not implemented by this runtime

The compiler half (`--preOptModules+=dataReconciliation`) already
worked; only the runtime half was missing. All 42 tests in
`openmodelica/dataReconciliation` pass with `--simCodeTarget=wasm-jit`,
and the debug logs, HTML reports and CSVs come out byte-identical to
the C runtime's — bar the run directory, the build's own version
string, and `ctime` being UTC here as everywhere else in the port.

The procedures live in `openmodelica_sim_meta::datarecon`, beside the
driver they hang off, so one copy serves the host-driven native run,
the in-wasm session driver and the wasip1 standalone command. Reports
go through `std::fs`: real files natively and under wasip1, the VFS in
the browser.

Model side: where C generates `data_function` / `setc_function` /
`setb_function` to copy variables into three `simulationInfo` arrays,
the runtime addresses the `SimData` slots directly, as `-l` already
does for its inputs and outputs. Only the symbolic Jacobians `F` and
`H` need emitted code, and they reuse the `-l` column emitter, whose
column-major output is the layout `getJacobianMatrixF` assembles by
hand. `functionDAE` (`functionLocalKnownVars` + `allEquations`) is
emitted only for a model the extraction algorithm ran on — it is not
`functionODE` + `functionAlgebraics`, which for a `when` model would
evaluate the ODE half twice.

C prints the procedure's log lines after the run's success line but
runs it while the model is still alive, so they are buffered and
replayed past the teardown hook, which is where the caller's capture
split puts post-success output.

`MODEL_FNS` was exactly 32 entries, so `present_mask` becomes a `u64`.

+++ OMCompiler/Compiler/OpenModelica.rs/openmodelica_wasm_jit/src/sim_runtime_wasmtime.rs
@@ -1598,11 +1598,11 @@ pub fn build_inwasm_session(model: &SimModel) -> std::result::Result<InWasmSessi
         .ok_or_else(|| "CodegenWasmJit: runtime has no __indirect_function_table export")?;
     let n_slots = crate::model::INWASM_SLOT_NAMES.len() as u64;
     let fn_base = wts(table.grow(&mut store, n_slots, wasmtime::Ref::Func(None)))?;
-    let mut present_mask: u32 = 0;
+    let mut present_mask: u64 = 0;
     for (slot, name) in crate::model::INWASM_SLOT_NAMES.iter().enumerate() {
         if let Some(f) = instance.get_func(&mut store, name) {
             wts(table.set(&mut store, fn_base + slot as u64, wasmtime::Ref::Func(Some(f))))?;
-            present_mask |= 1 << slot;
+            present_mask |= 1u64 << slot;
         }
     }
 
@@ -1629,7 +1629,7 @@ pub fn build_inwasm_session(model: &SimModel) -> std::result::Result<InWasmSessi
         return Err("CodegenWasmJit: the runtime rejected the simulation flags".to_string());
     }
 
-    let start = wts(rt_inst.get_typed_func::<(u32, u32, u32, u32), i32>(&mut store, "rt_sim_start"))?;
+    let start = wts(rt_inst.get_typed_func::<(u32, u32, u32, u64), i32>(&mut store, "rt_sim_start"))?;
     let gf = |store: &mut Store, name: &'static str| wts(rt_inst.get_typed_func::<(), u32>(store, name));
     // Assembled before the run starts so that an initialization `assert()`, which
     // traps out of `rt_sim_start`, is decoded through the same `SimEngine` as one
